+++ R/FSA-defunct.R
@@ -1,37 +1,58 @@
 #' @title Defunct functions.
 #' @description These functions were once part of FSA but have now been removed.
+#' @param \dots Additional arguments to pass through.
 #' @name FSA-defunct
-#' @keywords internal
-NULL
 
+#' @rdname FSA-defunct
+#' @export
+bootCase <- function(...) {
+  .Defunct(msg="'bootCase' has been removed; use 'car::Boot' with 'match='case'' (the default) for the same functionality.")
+}
+
+#' @rdname FSA-defunct
+#' @export
 chooseColors <- function(...) {
   .Defunct(msg="'chooseColors' has been removed; use one of many other available resources to choose appropriate colors.")
 }
 
+#' @rdname FSA-defunct
+#' @export
 compIntercepts <- function(...) {
   .Defunct(msg="'compIntercepts' has been removed; use 'emmeans' from the 'emmeans' package as described in the fishR blog post of 12-May-2021.")
 }
 
+#' @rdname FSA-defunct
+#' @export
 compSlopes <- function(...) {
   .Defunct(msg="'compSlopes' has been removed; use 'emtrends' from the 'emmeans' package as described in the fishR blog post of 11-May-2021.")
 }
 
+#' @rdname FSA-defunct
+#' @export
 diags <- function(...) {
   .Defunct(msg="'diags' has been removed (to 'FSAmisc' on GitHub).")
 }
 
+#' @rdname FSA-defunct
+#' @export
 hoCoef <- function(...) {
   .Defunct(msg="'hoCoef' has been removed (to 'FSAmisc' on GitHub).")
 }
 
+#' @rdname FSA-defunct
+#' @export
 mapvalues <- function(...) {
   .Defunct(msg="'FSA::mapvalues' has been removed; use 'plyr::mapvalues' instead for the same functionality.")
 }
 
+#' @rdname FSA-defunct
+#' @export
 plotBinResp <- function(...) {
   .Defunct(msg="'plotBinResp' has been removed; use 'ggplot2' as described in the fishR blog post of 25-May-2021.")
 }
 
+#' @rdname FSA-defunct
+#' @export
 Subset <- function(...) {
   .Defunct(msg="'Subset' has been removed; use 'subset' or 'dplyr::filter' followed by 'droplevels' for the same functionality.")
 }